requirements.txt
ibm-watson-machine-learning==1.0.327
ipywidgets==8.0.7
jupyter==1.0.0
matplotlib==3.7.2
numpy==1.24.2
pandas==1.5.3
plotly==5.15.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.11
rouge==1.0.1
scikit-learn==1.2.2
sentence-transformers==2.2.2
streamlit==1.24.1
safetensors==0.3.1
googletrans==3.1.0a0
langchain==0.0.321
pymysql==1.1.0
cryptography==41.0.5
